requirements.txt
openai==0.27.0
tweepy==4.15.0
lancedb==0.2.5
sentence-transformers==2.2.0
langchain==0.0.118
python-dotenv==1.0.0
retrying==1.3.3
